+++ src/siri/db/time.c
@@ -60,6 +60,6 @@ uint32_t siridb_time_in_seconds(siridb_t * siridb, int64_t ts)
 
 uint64_t siridb_time_now(siridb_t * siridb, struct timespec now)
 {
-    return now.tv_sec * siridb->time->factor +
-        now.tv_nsec * (siridb->time->factor / 1000000000.0);
+    return now.tv_sec * (uint64_t)siridb->time->factor +
+        now.tv_nsec * ((uint64_t)siridb->time->factor / 1000000000.0);
 }
